refactor(telegram): split collector into mixins
