+++ src/telegram/collector_mixins/__init__.py
@@ -1,0 +1,13 @@
+"""Collector composition mixins (#1137)."""
+
+from src.telegram.collector_mixins.cancellation import CancellationMixin
+from src.telegram.collector_mixins.collection import CollectionMixin
+from src.telegram.collector_mixins.stats import StatsMixin
+from src.telegram.collector_mixins.stream import StreamMixin
+
+__all__ = [
+    "CancellationMixin",
+    "CollectionMixin",
+    "StatsMixin",
+    "StreamMixin",
+]
